+Add character view
Added a view to add a character.
Wanted to adjust the ability scores based on
the race and class, but decided against in
because it gives unexpected results. The user
can rather keep the improvements in mind when
creating the character and set the scores as
desired. Keeps things a little simpler too.