+Character health view
Added view to retrieve, upate, and adjust a
character's health.
The GET, PUT, and PATCH work as expected. The
POST uses the AbilityScoreHealthMixin methods
to adjust the character's hp by the requested
amount. Adjusting the character's max HP will
has the option of adding the constitution
modifier, and willl also increase the
character's current HP proportionately.